pyproject.toml
[project]
name = "latterdns"
version = "0.1.0"
description = "A lightweight DNS proxy designed to bypass DNS spoofing by exploiting timing differences between fake and legitimate packets"
readme = "README.md"
requires-python = ">=3.14"
authors = [
    { name = "Amir Hosseini Nasab", email = "awmirhn@gmail.com" }
]
license = { text = "MIT" }
keywords = ["dns", "proxy", "censorship", "gfw", "anti-censorship"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "Topic :: Internet :: Name Service (DNS)",
    "Topic :: System :: Networking",
]
dependencies = [
    "click>=8.3.1",
    "dnspython>=2.8.0",
]

[project.scripts]
latterdns = "latterdns:main"

[build-system]
requires = ["uv_build>=0.9.13,<0.10.0"]
build-backend = "uv_build"

[project.urls]
Homepage = "https://github.com/itsamirhn/latterdns"
Repository = "https://github.com/itsamirhn/latterdns"
Issues = "https://github.com/itsamirhn/latterdns/issues"

[dependency-groups]
dev = [
    "ruff>=0.14.6",
]

[tool.ruff]
target-version = "py314"
line-length = 120

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "Q",      # flake8-quotes
    "RET",    # flake8-return
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate
    "PL",     # pylint
    "RUF",    # ruff-specific rules
    "PERF",   # perflint
    "FURB",   # refurb
    "LOG",    # flake8-logging
    "S",      # flake8-bandit (security)
    "A",      # flake8-builtins
    "COM",    # flake8-commas
    "DTZ",    # flake8-datetimez
    "T20",    # flake8-print
    "ARG",    # flake8-unused-arguments
]
ignore = [
    "COM812", # trailing comma (conflicts with formatter)
    "ISC001", # single line implicit string concatenation (conflicts with formatter)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.ruff.lint.isort]
known-first-party = ["latterdns"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
